refactor(whispering): align the os contract on fuji's platform/types.ts
fuji already established a single shared platform/types.ts holding every seam
contract. The os seam landed with a per-seam os.types.ts, which is a third
convention. Rename it to types.ts (with the shared-contract header and the
"stay free of @tauri-apps imports" rule fuji documents) so both platform-DI
apps share one layout. The contract stays seam-owned and neutral, so neither
impl depends on the other.
